These changes reduce the total number of compilation warnings.

+++ drivers/mbox/mbox.c
@@ -12,7 +12,7 @@ volatile uint32_t __attribute__((aligned(16))) mbox[36];
 
 int mbox_call(uint8_t channel)
 {
-  uint32_t r = (((uint32_t)(&mbox) & ~MBOX_MAIL_CHANNEL_MASK) | (channel & MBOX_MAIL_CHANNEL_MASK));
+  uint64_t r = (((uint64_t)(&mbox) & ~MBOX_MAIL_CHANNEL_MASK) | (channel & MBOX_MAIL_CHANNEL_MASK));
   // wait until we can talk to the VC
   while ( mmio_read(MBOX_STATUS) & MBOX_FULL ) { }
   // send our message to property channel and wait for the response
